# Projet_AMT/spring-server/src/main/resources/api-spec.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Gamification Platform API
  description: |
    **This is the documentation of the AMT Gamification Platform API**
host: localhost:8090
basePath: /api
schemes:
  - http
  - https
paths:
  /applications:
    get:
      description: "retrieve all applications"
      operationId: getAllApplications
      produces:
        - application/json
      responses:
        200:
          description:  "List all registered applications"
          schema:
            title: "Application summary"
            type: array
            items:
              $ref: '#/definitions/ApplicationSummary'
    post:
      description: "register a new application"
      operationId: postApplication
      consumes:
        - application/json
      parameters:
        - name: body
          in: body
          description: "The info required to register an application"
          schema:
            $ref: '#/definitions/ApplicationRegistration'
          required: true
      responses:
        201:
          description: "Register a new application"
        404:
          description: "Application not found"
        405:
          description: "Invalid input"
    put:
      description: "update an existing application"
      operationId: updateApplication
      consumes:
        - application/json
      parameters:
        - name: body
          in: body
          description: "The info required to update an application"
          schema:
            $ref: '#/definitions/ApplicationUpdate'
          required: true
      responses:
        201:
          description: "Update an existing application"
        404:
          description: "Application not found"
        405:
          description: "Invalid input"
  /auth:
    post:
      operationId: authenticateApplicationAndGetToken
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: "The info required to authenticate an application"
          schema:
            $ref: '#/definitions/Credentials'
          required: true
      responses:
        200:
          description: "Authentication sucessful"
          schema:
            $ref: '#/definitions/Token'
        401:
          description: "Authentication failed"
...
  /badges/{uuid}:
    get:
      description: "Retrieve badges by application uuid"
      operationId: findApplicationBadges
      parameters:
        - name: uuid
          in: path
          description: "uuid de l'application à trouver"
          required: true
          type: string
      produces:
        - application/json
      responses:
        200:
          description: "List all badges applications"
          schema:
            title: "Application Badges summary"
            type: array
            items:
              $ref: '#/definitions/ApplicationsBadgesSummary'
        404:
          description: "Application not found"
        405:
          description: "Invalid input"
    post:
      description: "add new badge"
      operationId: postBadge
      consumes:
        - application/json
      parameters:
        - name: uuid
          in: path
          description: "uuid of the application to fetch"
          required: true
          type: string
        - name: body
          in: body
          description: "The info required to register an application's badge"
          schema:
            $ref: '#/definitions/RegistrationBadge'
          required: true
      responses:
        201:
          description: "Register a new badge"
        405:
          description: "Invalid input"
    put:
      description: "update a badge"
      operationId: updateBadge
      consumes:
        - application/json
      parameters:
        - name: uuid
          in: path
          description: "uuid of the application to fetch"
          required: true
          type: string
        - name: body
          in: body
          description: "The info required to update an application's badge"
          schema:
            $ref: '#/definitions/UpdateBadge'
          required: true
      responses:
        201:
          description: "Update a badge"
        405:
          description: "Invalid input"
...
definitions:
  ApplicationSummary:
    type: object
    properties:
      applicationName:
        type: string
      applicationKey:
        type: string
  ApplicationRegistration:
    type: object
    properties:
      applicationName:
        type: string
      applicationKeyUUID:
        type: string
      applicationSecretUUID:
        type: string
  ApplicationUpdate:
    type: object
    properties:
      applicationNewName:
        type: string
      applicationKey:
        type: string
      applicationSecretUUID:
        type: string
  Credentials:
    type: object
    properties:
      applicationName:
        type: string
      password:
        type: string
  Token:
    type: object
    properties:
      applicationName:
        type: string
...
  RegistrationBadge:
    type: object
    properties:
      badgeName:
        type: string
      applicationSecret:
        type: string
  UpdateBadge:
    type: object
    properties:
      oldBadgeName:
        type: string
      badgeName:
        type: string
      applicationSecret:
        type: string
  ApplicationsBadgesSummary:
    type: object
    properties:
      badgesName:
        type: string
...